Fix client/server issues

+++ src/shared/src/buffer.cpp
@@ -2,11 +2,11 @@
 
 #include <cassert>
 
-size_t Reader::sizeLeft() const {
+int64_t Reader::sizeLeft() const {
     return size - index;
 }
 
-Reader::Reader(const uint8_t *data, size_t size) : data(data), size(size) { }
+Reader::Reader(const uint8_t *data, int64_t size) : data(data), size(size) { }
 
 Reader::~Reader() {
     if (size != index)
